requirements.txt
yt-dlp==2023.10.13
whisper==1.1.10
googletrans==4.0.0-rc1
ffmpeg-python==0.2.0